refactor: centralize request logic
